server structure overhauled with minor typos in client corrected
