+++ server/package.json
@@ -27,6 +27,8 @@
   },
   "devDependencies": {
     "@types/node": "^20.12.4",
-    "nodemon": "^3.1.0"
+    "jest": "^29.7.0",
+    "nodemon": "^3.1.0",
+    "supertest": "^7.0.0"
   }
 }
